requirements.txt
pyautogui
Pillow
pyyaml
# For the coordinate finder:
keyboard